增加获取cookie与持久化存储活动

+++ src/fecth.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
+import os.path
 import requests
 import schedule
 import time
@@ -7,9 +8,12 @@ import random
 import send_mes
 from datetime import datetime
 import sys
+import get_cookies
 
 # 用于存储活动名称的集合
 existing_activities_id = set()
+# 用于更新存储数据
+new_activities_id = set()
 
 headers = {
     'User-Agent': '',
@@ -100,37 +104,76 @@ def fetch_activities():
 
     return True  # 返回True表示成功
 
+def send_error(content):
+    send_mes.send(content)
 
 def init():
     now = datetime.now()
-    with open('log.txt', 'w', encoding='utf-8') as file:
-        file.write(f"内容：日志\n创建时间：{now}\n")
+    if not os.path.exists('log.txt'):
+        with open('log.txt', 'w', encoding='utf-8') as file:
+            file.write(f"内容：日志\n创建时间：{now}\n")
+    if not os.path.exists('activites.txt'):
+        with open('activites.txt','w',encoding='utf-8') as file:
+            file.write(f"内容：活动列表\n创建世界：{now}\n")
+    else:
+        with open('activites.txt','r',encoding='utf-8') as file:
+            lines = file.readlines()
+            if len(lines) >= 3:  # 确保有足够的行
+                integer_line = lines[2].strip()  # 获取第三行的整数数据
+                integers = map(int, integer_line.split())  # 将字符串分割并转换为整数
+                existing_activities_id.update(integers)  # 将整数添加到集合中
+    P=get_cookies.get_cookies()
+    if P:
+        cookies["PHPSESSID"]=P
+        print(f"新获取的cookie是:{P}")
+    else:
+        log("cookie获取错误")
+        send_error(["cookie获取错误，请尽快检查程序"])
+        sys.exit()
 
-def send_error(content):
-    send_mes.send(content)
+
+def updata_activites():
+    with open('activites.txt','a',encoding='utf-8') as file:
+        for item in new_activities_id:
+            file.write(f' {item}')
+
 
 def job():
     reslut=fetch_activities()
     if not reslut:  #获取失败，强制退出程序
-        log("异常退出")
-        send_error(["异常退出，请检查日志文件解决问题"])
-        sys.exit()
-
+        #尝试获取新cookie
+        res=get_cookies.get_cookies()
+        if res:
+            cookies['PHPSESSID']=res
+            #检查是否可以了
+            if not fetch_activities():
+                # 不是cookie错误，其余问题
+                log("异常退出")
+                send_error(["异常退出，请检查日志文件解决问题"])
+                updata_activites()  # 更新数据，以便下次启动重复发送数据
+                sys.exit()
+        else:
+            log("cookie获取错误")
+            send_error(["cookie获取错误，请尽快检查程序"])
+            updata_activites()  # 更新数据，以便下次启动重复发送数据
+            sys.exit()
 
 # 启动爬虫
 if __name__ == '__main__':
-    init()  # 初始化日志
+    init()  # 初始化
     cnt=1
     limit=168   #7天退出一次
 
     # 定时任务设置
-    schedule.every(6).hours.do(job)  # 每6个小时执行一次
+    schedule.every(2).hours.do(job)  # 每2个小时执行一次
 
     while cnt<=limit:
         schedule.run_pending()
         time.sleep(3600)
         cnt+=1
+        if cnt<=20: # 前5次测试程序运行效果
+            send_mes.send([f'程序测试中，第{cnt}次'])
         print("检查...")
     log("正常退出")
     send_error(["正常执行结束，若有需要请自行重新启动"])
-
+    updata_activites()  # 更新数据，以便下次启动重复发送数据
